# docs/openapi.yaml
openapi: 3.1.0
info:
  title: RiceText 论坛与小说富文本 API
  version: 0.1.0
  description: 文档、章节、用户互动、附件权益和投票均使用共享契约，并由 SQLite 服务持久化实现。
servers:
  - url: http://localhost:8787
    description: 本地开发 API
components:
  securitySchemes:
    cookieSession:
      type: apiKey
      in: cookie
      name: ricetext_session
      description: OIDC 登录后由 Worker 设置的 HttpOnly session cookie。
...
paths:
...
  /api/documents/{documentId}/comments/{anchorId}:
    get:
      operationId: getCommentThread
      tags:
        - 间贴
      summary: 读取间贴回复树
      description: |-
        只对根回复做游标分页，返回根节点的完整后代；sort=score 或 newest。

        生产使用 HttpOnly session cookie；x-user-id 仅供显式启用的本地 demo 模式。
      parameters:
        - name: documentId
          in: path
          required: true
          description: 路径参数 documentId
          schema:
            type: string
            minLength: 1
            maxLength: 128
            pattern: ^[A-Za-z0-9][A-Za-z0-9_-]*$
          example: demo-post
        - name: anchorId
          in: path
          required: true
          description: 路径参数 anchorId
          schema:
            type: string
            minLength: 1
            maxLength: 128
            pattern: ^[A-Za-z0-9][A-Za-z0-9_-]*$
        - name: cursor
          in: query
          required: false
          description: 查询参数 cursor
          schema:
            type: string
            minLength: 1
            maxLength: 256
        - name: limit
          in: query
          required: false
          description: 查询参数 limit
          schema:
            default: 20
            type: integer
            minimum: 1
            maximum: 100
          example: 20
        - name: sort
          in: query
          required: false
          description: 查询参数 sort
          schema:
            default: score
            type: string
            enum:
              - score
              - newest
        - name: x-user-id
          in: header
          required: false
          description: 仅限本地 demo：论坛身份 ID author、reader 或 moderator。preview/production 会忽略该请求头。
          schema:
            type: string
            enum:
              - author
              - reader
              - moderator
            default: reader
          example: reader
      security:
        - cookieSession: []
      responses:
        "200":
          description: 锚点的树状回复及当前身份投票。
          content:
            application/json:
              schema:
                type: object
                properties:
                  documentId:
                    type: string
                    minLength: 1
                    maxLength: 128
                    pattern: ^[A-Za-z0-9][A-Za-z0-9_-]*$
                  anchorId:
                    type: string
                    minLength: 1
                    maxLength: 128
                    pattern: ^[A-Za-z0-9][A-Za-z0-9_-]*$
                  archived:
                    type: boolean
                  total:
                    type: integer
                    minimum: 0
                    maximum: 9007199254740991
                  items:
                    type: array
                    items:
                      type: object
                      description: 递归节点，完整结构由 @ricetext/contracts 的 Zod 契约校验。
                      additionalProperties: true
                  pageInfo:
                    type: object
                    properties:
                      nextCursor:
                        anyOf:
                          - type: string
                          - type: "null"
                    required:
                      - nextCursor
                    additionalProperties: false
                required:
                  - documentId
                  - anchorId
                  - archived
                  - total
                  - items
                  - pageInfo
                additionalProperties: false
        "401":
          description: 未登录或会话已过期。
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        minLength: 1
                      message:
                        type: string
                        minLength: 1
                      details:
                        type: object
                        propertyNames:
                          type: string
                        additionalProperties:
                          type: object
                          description: 递归节点，完整结构由 @ricetext/contracts 的 Zod 契约校验。
                          additionalProperties: true
                    required:
                      - code
                      - message
                    additionalProperties: false
                required:
                  - error
                additionalProperties: false
        "404":
          description: 文档不存在。
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        minLength: 1
                      message:
                        type: string
                        minLength: 1
                      details:
                        type: object
                        propertyNames:
                          type: string
                        additionalProperties:
                          type: object
                          description: 递归节点，完整结构由 @ricetext/contracts 的 Zod 契约校验。
                          additionalProperties: true
                    required:
                      - code
                      - message
                    additionalProperties: false
                required:
                  - error
                additionalProperties: false
              example:
                error:
                  code: EXAMPLE_ERROR
                  message: 文档不存在。
  /api/documents/{documentId}/comments/{anchorId}/replies:
    post:
      operationId: createCommentReply
      tags:
        - 间贴
      summary: 新增间贴回复
      description: |-
        所有论坛身份可回复。parentId 为 null 创建根回复，否则创建楼中楼。

        生产使用 HttpOnly session cookie；x-user-id 仅供显式启用的本地 demo 模式。
      parameters:
        - name: documentId
          in: path
          required: true
          description: 路径参数 documentId
          schema:
            type: string
            minLength: 1
            maxLength: 128
            pattern: ^[A-Za-z0-9][A-Za-z0-9_-]*$
          example: demo-post
        - name: anchorId
          in: path
          required: true
          description: 路径参数 anchorId
          schema:
            type: string
            minLength: 1
            maxLength: 128
            pattern: ^[A-Za-z0-9][A-Za-z0-9_-]*$
        - name: x-user-id
          in: header
          required: false
          description: 仅限本地 demo：论坛身份 ID author、reader 或 moderator。preview/production 会忽略该请求头。
          schema:
            type: string
            enum:
              - author
              - reader
              - moderator
            default: reader
          example: reader
      security:
        - cookieSession: []
      requestBody:
        required: true
        description: 请求字段由共享 Zod 契约校验。
        content:
          application/json:
            schema:
              type: object
              properties:
                parentId:
                  default: null
                  anyOf:
                    - type: string
                      minLength: 1
                      maxLength: 128
                      pattern: ^[A-Za-z0-9][A-Za-z0-9_-]*$
                    - type: "null"
                body:
                  type: string
                  minLength: 1
                  maxLength: 10000
              required:
                - parentId
                - body
              additionalProperties: false
      responses:
        "201":
          description: 新回复；children 初始为空。
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    minLength: 1
                    maxLength: 128
                    pattern: ^[A-Za-z0-9][A-Za-z0-9_-]*$
                  parentId:
                    anyOf:
                      - type: string
                        minLength: 1
                        maxLength: 128
                        pattern: ^[A-Za-z0-9][A-Za-z0-9_-]*$
                      - type: "null"
                  author:
                    type: object
                    properties:
                      id:
                        type: string
                        minLength: 1
                        maxLength: 128
                        pattern: ^[A-Za-z0-9][A-Za-z0-9_-]*$
                      name:
                        type: string
                      role:
                        type: string
                        enum:
                          - author
                          - reader
                          - moderator
                      avatar:
                        type: string
                      coins:
                        type: integer
                        minimum: -9007199254740991
                        maximum: 9007199254740991
                      replied:
                        type: boolean
                    required:
                      - id
                      - name
                      - role
                      - avatar
                      - coins
                      - replied
                    additionalProperties: false
                  body:
                    type: string
                  score:
                    type: integer
                    minimum: -9007199254740991
                    maximum: 9007199254740991
                  viewerVote:
                    anyOf:
                      - type: number
                        const: -1
                      - type: number
                        const: 0
                      - type: number
                        const: 1
                  upvotes:
                    type: integer
                    minimum: 0
                    maximum: 9007199254740991
                  downvotes:
                    type: integer
                    minimum: 0
                    maximum: 9007199254740991
                  myVote:
                    anyOf:
                      - type: number
                        const: -1
                      - type: number
                        const: 0
                      - type: number
                        const: 1
                  createdAt:
                    type: string
                    format: date-time
                    pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z|([+-](?:[01]\d|2[0-3]):[0-5]\d)))$
                  children:
                    type: array
                    items:
                      type: object
                      description: 递归节点，完整结构由 @ricetext/contracts 的 Zod 契约校验。
                      additionalProperties: true
                required:
                  - id
                  - parentId
                  - author
                  - body
                  - score
                  - viewerVote
                  - upvotes
                  - downvotes
                  - myVote
                  - createdAt
                  - children
                additionalProperties: false
        "401":
          description: 未登录或会话已过期。
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        minLength: 1
                      message:
                        type: string
                        minLength: 1
                      details:
                        type: object
                        propertyNames:
                          type: string
                        additionalProperties:
                          type: object
                          description: 递归节点，完整结构由 @ricetext/contracts 的 Zod 契约校验。
                          additionalProperties: true
                    required:
                      - code
                      - message
                    additionalProperties: false
                required:
                  - error
                additionalProperties: false
        "404":
          description: 文档或父回复不存在。
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        minLength: 1
                      message:
                        type: string
                        minLength: 1
                      details:
                        type: object
                        propertyNames:
                          type: string
                        additionalProperties:
                          type: object
                          description: 递归节点，完整结构由 @ricetext/contracts 的 Zod 契约校验。
                          additionalProperties: true
                    required:
                      - code
                      - message
                    additionalProperties: false
                required:
                  - error
                additionalProperties: false
              example:
                error:
                  code: EXAMPLE_ERROR
                  message: 文档或父回复不存在。
        "409":
          description: 锚点已随正文删除并归档。
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        minLength: 1
                      message:
                        type: string
                        minLength: 1
                      details:
                        type: object
                        propertyNames:
                          type: string
                        additionalProperties:
                          type: object
                          description: 递归节点，完整结构由 @ricetext/contracts 的 Zod 契约校验。
                          additionalProperties: true
                    required:
                      - code
                      - message
                    additionalProperties: false
                required:
                  - error
                additionalProperties: false
              example:
                error:
                  code: EXAMPLE_ERROR
                  message: 锚点已随正文删除并归档。
  /api/comments/replies/{replyId}/vote:
    put:
      operationId: voteComment
      tags:
        - 间贴
      summary: 赞、踩或撤销
      description: |-
        value=1 点赞、-1 点踩、0 撤销；每个身份对每条回复只有一票。

        生产使用 HttpOnly session cookie；x-user-id 仅供显式启用的本地 demo 模式。
      parameters:
        - name: replyId
          in: path
          required: true
          description: 路径参数 replyId
          schema:
            type: string
            minLength: 1
            maxLength: 128
            pattern: ^[A-Za-z0-9][A-Za-z0-9_-]*$
        - name: x-user-id
          in: header
          required: false
          description: 仅限本地 demo：论坛身份 ID author、reader 或 moderator。preview/production 会忽略该请求头。
          schema:
            type: string
            enum:
              - author
              - reader
              - moderator
            default: reader
          example: reader
      security:
        - cookieSession: []
      requestBody:
        required: true
        description: 请求字段由共享 Zod 契约校验。
        content:
          application/json:
            schema:
              type: object
              properties:
                value:
                  anyOf:
                    - type: number
                      const: -1
                    - type: number
                      const: 0
                    - type: number
                      const: 1
              required:
                - value
              additionalProperties: false
      responses:
        "200":
          description: 更新后的净赞数、赞踩计数和当前投票。
          content:
            application/json:
              schema:
                type: object
                properties:
                  score:
                    type: integer
                    minimum: -9007199254740991
                    maximum: 9007199254740991
                  viewerVote:
                    anyOf:
                      - type: number
                        const: -1
                      - type: number
                        const: 0
                      - type: number
                        const: 1
                  upvotes:
                    type: integer
                    minimum: 0
                    maximum: 9007199254740991
                  downvotes:
                    type: integer
                    minimum: 0
                    maximum: 9007199254740991
                  myVote:
                    anyOf:
                      - type: number
                        const: -1
                      - type: number
                        const: 0
                      - type: number
                        const: 1
                required:
                  - score
                  - viewerVote
                  - upvotes
                  - downvotes
                  - myVote
                additionalProperties: false
        "401":
          description: 未登录或会话已过期。
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        minLength: 1
                      message:
                        type: string
                        minLength: 1
                      details:
                        type: object
                        propertyNames:
                          type: string
                        additionalProperties:
                          type: object
                          description: 递归节点，完整结构由 @ricetext/contracts 的 Zod 契约校验。
                          additionalProperties: true
                    required:
                      - code
                      - message
                    additionalProperties: false
                required:
                  - error
                additionalProperties: false
        "404":
          description: 回复不存在。
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        minLength: 1
                      message:
                        type: string
                        minLength: 1
                      details:
                        type: object
                        propertyNames:
                          type: string
                        additionalProperties:
                          type: object
                          description: 递归节点，完整结构由 @ricetext/contracts 的 Zod 契约校验。
                          additionalProperties: true
                    required:
                      - code
                      - message
                    additionalProperties: false
                required:
                  - error
                additionalProperties: false
              example:
                error:
                  code: EXAMPLE_ERROR
                  message: 回复不存在。
...